test: split app behavior coverage
